+++ src/main/java/asedi/services/LocalService.java
@@ -8,14 +8,14 @@ import asedi.model.Local;
 public class LocalService {
     // TODO: Implementar conexión real con la API
     
-    public List<Local> obtenerLocalesPorPlaza(int plazaId) {
+    public List<Local> obtenerLocalesPorPlaza(Long plazaId) {
         // Simulación de datos para pruebas
         List<Local> locales = new ArrayList<>();
         
         // Datos de ejemplo
-        if (plazaId == 1) {
-            Local local1 = new Local(1, 1, "Restaurante La Esquina", "Comida mexicana tradicional", "Local 12", "09:00", "22:00", "Restaurante");
-            Local local2 = new Local(2, 1, "Café Aromas", "Café de especialidad y repostería", "Local 15", "07:00", "21:00", "Cafetería");
+        if (plazaId == 1L) {
+            Local local1 = new Local(1L, 1L, "Restaurante La Esquina", "Comida mexicana tradicional", "Local 12", "09:00", "22:00", "Restaurante");
+            Local local2 = new Local(2L, 1L, "Café Aromas", "Café de especialidad y repostería", "Local 15", "07:00", "21:00", "Cafetería");
             locales.add(local1);
             locales.add(local2);
         }
@@ -35,7 +35,7 @@ public class LocalService {
         return true;
     }
     
-    public boolean eliminarLocal(int localId) {
+    public boolean eliminarLocal(Long localId) {
         // TODO: Implementar lógica para eliminar el local
         System.out.println("Eliminando local con ID: " + localId);
         return true;
